requirements.txt
# to install
# $ pip install -r requirements.txt
# to upgrade all
# $ pip install -r requirements.txt --upgrade
# to remove all
# $ pip freeze | xargs pip uninstall -y
# to list outdated packages
# pip list --outdated --format=freeze

dnspython==1.16.0
unificontrol==0.2.7
